fix(yank): address second bugbot round — dedup, validation, UX
- Move manifestOwnedByUser to api/lib/auth-helpers and export it,
  removing the duplicate in yank.js
- Tighten PKG_RE: allow @scope/name but block bare path traversal
  sequences (was /^[\w.@/-]+$/, now /^(?:@[\w.-]+\/)?[\w.+-]+$/)
- Sanitize 500 error responses — log to console, return generic
  error code without leaking internal exception messages
- Add Unyank confirmation step (Unyank? Yes / Cancel) matching Yank UX
